# internal/swagger/swagger.yaml
basePath: /
definitions:
...
  controller.SimpleResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: Swagger documentation for Tinyauth's API.
  license:
    name: AGPL-3.0
    url: https://github.com/tinyauthapp/tinyauth/blob/main/LICENSE
  title: Tinyauth API
  version: development
paths:
...
  /api/auth/caddy:
    get:
      description: Forward-Auth Proxy Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: Proxy
      tags:
      - forward-auth
  /api/auth/envoy:
    delete:
      description: Forward-Auth Proxy Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: Proxy
      tags:
      - forward-auth
    get:
      description: Forward-Auth Proxy Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: Proxy
      tags:
      - forward-auth
    head:
      description: Forward-Auth Proxy Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: Proxy
      tags:
      - forward-auth
    options:
      description: Forward-Auth Proxy Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: Proxy
      tags:
      - forward-auth
    patch:
      description: Forward-Auth Proxy Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: Proxy
      tags:
      - forward-auth
    post:
      description: Forward-Auth Proxy Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: Proxy
      tags:
      - forward-auth
    put:
      description: Forward-Auth Proxy Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: Proxy
      tags:
      - forward-auth
  /api/auth/nginx:
    get:
      description: Forward-Auth Proxy Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: Proxy
      tags:
      - forward-auth
  /api/auth/traefik:
    get:
      description: Forward-Auth Proxy Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.SimpleResponse'
      summary: Proxy
      tags:
      - forward-auth
...
swagger: "2.0"